Map updated
Updated the ocean map to another page
